+++ src/modules/score.c
@@ -161,8 +161,8 @@ void score(void)
             if (f->num_total != 0) {
                 fprintf(scoreFP, "%8d (%8d/%4.2f%%/%5.2f) %30.30s (%3.3s) %5s (%3d)\n",
                     f->score, f->score - average_score_of_age(f->age, f->age / 24 + 1),
-                    ((float)f->score / (float)allscores) * 100.0,
-                    (float)f->score / f->num_total,
+                    ((double)f->score / allscores) * 100,
+                    (double)f->score / f->num_total,
                     f->name, LOC(default_locale, rc_name_s(f->race, NAME_SINGULAR)), factionid(f),
                     f->age);
             }
